% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Change of data form}
\usage{
change(data)
}
\arguments{
\item{original}{data}
}
\value{
Reshaped data
}
\description{
Taking input as original data and output as another form that is easy to handle.
}
\examples{
   change(data)
}
