% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{reform}
\alias{reform}
\title{wide form data to long form}
\usage{
reform(data)
}
\arguments{
\item{changed}{data}
}
\value{
reformatted data
}
\description{
Taking data resulted from the function 'change' and produce a data frame that can be used for plotting the diagram.
}
\examples{
   reform(data)
}
