% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyplot.R
\name{sankeyplot}
\alias{sankeyplot}
\title{Sankey Diagram}
\usage{
sankeyplot(data, a, b)
}
\arguments{
\item{data, }{vertical white line (0 or 1), horizontal white line (0 or 1)}
}
\value{
diagram
}
\description{
Plotting versions of Sankey diagrams
}
\examples{
   reform(data, 1, 0)
}
